+++ pipeline/include/cvs/pipeline/iexecutionNode.hpp
