+++ pipeline/include/cvs/pipeline/imoduleManager.hpp
