+++ pipeline/include/cvs/pipeline/tbb/tbbHelpers.hpp
@@ -4,19 +4,19 @@
 #include <cvs/common/factory.hpp>
 #include <cvs/logger/logging.hpp>
 #include <cvs/pipeline/ielement.hpp>
-#include <cvs/pipeline/iexecutiongraph.hpp>
-#include <cvs/pipeline/registrationhelper.hpp>
+#include <cvs/pipeline/iexecutionGraph.hpp>
+#include <cvs/pipeline/registrationHelper.hpp>
 #include <cvs/pipeline/tbb/tbbBroadcastNode.hpp>
 #include <cvs/pipeline/tbb/tbbBufferNode.hpp>
+#include <cvs/pipeline/tbb/tbbContinueNode.hpp>
+#include <cvs/pipeline/tbb/tbbDefinitions.hpp>
+#include <cvs/pipeline/tbb/tbbFlowGraph.hpp>
+#include <cvs/pipeline/tbb/tbbFunctionNode.hpp>
+#include <cvs/pipeline/tbb/tbbJoinNode.hpp>
 #include <cvs/pipeline/tbb/tbbMultifunctionNode.hpp>
 #include <cvs/pipeline/tbb/tbbOverwriteNode.hpp>
-#include <cvs/pipeline/tbb/tbbcontinuenode.hpp>
-#include <cvs/pipeline/tbb/tbbdefinitions.hpp>
-#include <cvs/pipeline/tbb/tbbflowgraph.hpp>
-#include <cvs/pipeline/tbb/tbbfunctionnode.hpp>
-#include <cvs/pipeline/tbb/tbbjoinnode.hpp>
-#include <cvs/pipeline/tbb/tbbsourcenode.hpp>
-#include <cvs/pipeline/tbb/tbbsplitnode.hpp>
+#include <cvs/pipeline/tbb/tbbSourceNode.hpp>
+#include <cvs/pipeline/tbb/tbbSplitNode.hpp>
 
 #include <iostream>
 #include <type_traits>
