+++ tool/src/cvspipelinetool.cpp
@@ -1,4 +1,4 @@
-#include <cvs/pipeline/registrationhelper.hpp>
-#include <cvs/pipeline/tbb/tbbhelpers.hpp>
+#include <cvs/pipeline/registrationHelper.hpp>
+#include <cvs/pipeline/tbb/tbbHelpers.hpp>
 
 cvs::common::FactoryPtr<std::string> pipelineFactory() { return cvs::common::Factory<std::string>::defaultInstance(); }
